% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conapop}
\alias{conapop}
\title{Conapo 1990-2040.}
\format{

}
\source{
<https://www.gob.mx/conapo/documentos/reconstruccion-y-proyecciones-de-la-poblacion-de-los-municipios-de-mexico-1990-2040>
}
\usage{
conapop
}
\description{
Reconstruction and projections of the population of the municipalities of Mexico 1990-2040
dataset ...
}
\details{
Mid-year population figures are presented for 2,475 municipalities in Mexico, from 1990 to 2020, and population projections from 2021 to 2040.
}
\keyword{datasets}
