% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pop.R
\name{extract_pop}
\alias{extract_pop}
\title{extracts the population density at the locality level in hexagons of 400 meters resolution for the years 2023 and 2022}
\usage{
extract_pop(year, cve_edo, locality)
}
\arguments{
\item{year}{is the year dataset}

\item{cve_edo}{is the id of state.}

\item{locality}{is the target locality.}
}
\value{
a sf object
}
\description{
extracts the population density at the locality level in hexagons of 400 meters resolution for the years 2023 and 2022
}
\examples{
popmex::extract_pop(year = 2023, cve_edo = "31", locality = "Merida")
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}.
}
