% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hdx_pop2021}
\alias{hdx_pop2021}
\title{Population Density Mexico HDX data 2021}
\format{
## `hdx_pop2021`
A sf & data frame with 362028 rows and 2 columns:
\describe{
  \item{popolation}{Population Density for 400m H3 Hexagons}
  \item{geom}{geometry}
  ...
}
}
\source{
<https://data.humdata.org/dataset/kontur-population-dataset>
}
\usage{
hdx_pop2021
}
\description{
A subset of data from the HUMANITARIAN DATA EXCHANGE Population Density of Mexico
dataset ...
}
\details{
EPSG 4326...
}
\keyword{datasets}
