% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hdx_pop2022}
\alias{hdx_pop2022}
\title{Population Density Mexico HDX data 2022}
\format{
## `hdx_pop2022`
A sf & data frame with 463830 rows and 3 columns:
\describe{
  \item{h3}{id of hexagon}
  \item{popolation}{Population Density for 400m H3 Hexagons}
  \item{geom}{geometry}
  ...
}
}
\source{
<https://data.humdata.org/dataset/kontur-population-mexico?>
}
\usage{
hdx_pop2022
}
\description{
A subset of data from the HUMANITARIAN DATA EXCHANGE Population Density of Mexico
dataset ...
}
\keyword{datasets}
