% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop2000}
\alias{pop2000}
\title{population by municipality (ITER 2000).}
\format{
## `pop2000`
A data frame with 2443 rows and 4 columns:
\describe{
  \item{CVEGEO}{id of the municipality.}
  \item{pobtot}{population}
  \item{pobfem}{female population}
  \item{pobmas}{male population}
  ...
}
}
\source{
<https://www.inegi.org.mx/app/scitel/Default?ev=9>
}
\usage{
pop2000
}
\description{
a dataset of population by municipality of mexico.
dataset ...
}
\details{
The population data by total number and sex for each municipality were obtained from the dataset of territorial indicators known as ITER /(Indicadores Territoriales/), which is a database derived from the 2020 Population and Housing Census generated by Instituto Nacional de Estadística y Geografía (INEGI)
}
\keyword{datasets}
