% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop2010}
\alias{pop2010}
\title{population by municipality (ITER 2010).}
\format{
## `pop2010`
A data frame with 2456 rows and 44 columns:
\describe{
  \item{CVEGEO}{id of the municipality.}
  \item{pobtot}{population}
  \item{pobfem}{female population}
  \item{pobmas}{male population}
  \item{p_0a2}{population between 0 and 2 years of age}
  \item{p_0a2_f}{female population between 0 and 2 years of age}
  \item{p_0a2_m}{male population between 0 and 2 years of age}
  \item{p_3ymas}{population over 3 years of age}
  \item{p_3ymas_f}{female population over 3 years of age}
  \item{p_3ymas_m}{male population over 3 years of age}
  \item{p_5ymas}{population over 5 years of age}
  \item{p_5ymas_f}{female population over 5 years of age}
  \item{p_5ymas_m}{male population over 5 years of age}
  \item{p_12ymas}{population over 12 years of age}
  \item{p_12ymas_f}{female population over 12 years of age}
  \item{p_12ymas_m}{male population over 12 years of age}
  \item{p_15ymas}{population over 15 years of age}
  \item{p_15ymas_f}{female population over 15 years of age}
  \item{p_15ymas_m}{male population over 15 years of age}
  \item{p_18ymas}{population over 18 years of age}
  \item{p_18ymas_f}{female population over 18 years of age}
  \item{p_18ymas_m}{male population over 18 years of age}
  \item{p_3a5}{population between 3 and 5 years of age}
  \item{p_3a5_f}{female population between 3 and 5 years of age}
  \item{p_3a5_m}{male population between 3 and 5 years of age}
  \item{p_6a11}{population between 6 and 11 years of age}
  \item{p_6a11_f}{female population between 6 and 11 years of age}
  \item{p_6a11_m}{male population between 6 and 11 years of age}
  \item{p_8a14}{population between 8 and 14 years of age}
  \item{p_8a14_f}{female population between 8 and 14 years of age}
  \item{p_8a14_m}{male population between 8 and 14 years of age}
  \item{p_12a14}{population between 12 and 14 years of age}
  \item{p_12a14_f}{female population between 12 and 14 years of age}
  \item{p_12a14_m}{male population between 12 and 14 years of age}
  \item{p_15a17}{population between 15 and 17 years of age}
  \item{p_15a17_f}{female population between 15 and 17 years of age}
  \item{p_15a17_m}{male population between 15 and 17 years of age}
  \item{p_18a24}{population between 18 and 24 years of age}
  \item{p_18a24_f}{female population between 18 and 24 years of age}
  \item{p_18a24_m}{male population between 18 and 24 years of age}
  \item{p_15a49_f}{female population between 15 and 49 years of age}
  \item{p_60ymas}{population over 60 years of age}
  \item{p_60ymas_f}{female population over 60 years of age}
  \item{p_60ymas_m}{male population over 60 years of age}
  ...
}
}
\source{
<https://www.inegi.org.mx/app/scitel/Default?ev=9>
}
\usage{
pop2010
}
\description{
a dataset of population by municipality of mexico.
dataset ...
}
\details{
The population data by total number and sex for each municipality were obtained from the dataset of territorial indicators known as ITER /(Indicadores Territoriales/), which is a database derived from the 2020 Population and Housing Census generated by Instituto Nacional de Estadística y Geografía (INEGI)
}
\keyword{datasets}
